[NEW] Add support for file size weight selection.
Using the traditiona max_filesize and min_filesize has its limits when you want a 
large number of small files and only a few big files.  The new size_weight option
allow more fine grain selection of the size of needed as the distribution of such
file.  The selection is not stick but it give fairly accurate.  There are no 
limits to the number of size_weight that can be used.

Here a small example on how to use.

[filesystem]
	location=/mnt
	size_weight 4096 60
	size_weight 1048576 30
	size_weight 10485760 9
	size_weight 104857600 1
[end]