[NEW] Add fsync equivalent for all write operations.
Added new write_fsync_weight, create_fsync_weight and
append_fsync_weight which are identical to the non-fsync operations
with the one obvious difference.  This allows greater control on how
the write operations are handled.  The old fsync_file option in the
profiles is as of now deprecated and will be removed in the future.